added text class for drawing strings with the specified font- discovered possible driver bug (java returned -1073740940?)

+++ src/dev/theskidster/xjge2/ui/Glyph.java
@@ -1,7 +1,7 @@
 package dev.theskidster.xjge2.ui;
 
 import dev.theskidster.xjge2.graphics.Color;
-import org.joml.Vector3i;
+import org.joml.Vector3f;
 
 /**
  * @author J Hoffman
@@ -12,10 +12,10 @@ final class Glyph {
 
     char c;
     
-    Vector3i position;
+    Vector3f position;
     Color color;
     
-    Glyph(char c, Vector3i position, Color color) {
+    Glyph(char c, Vector3f position, Color color) {
         this.c        = c;
         this.position = position;
         this.color    = color;
